rename the file to ...Slice.tsx
